+++ R/AllGenerics.R
@@ -513,7 +513,8 @@ setGeneric("chromPeakData<-", function(object, value)
 #'
 #' ## Read a file with DDA LC-MS/MS data
 #' library(MsExperiment)
-#' fl <- system.file("TripleTOF-SWATH/PestMix1_DDA.mzML", package = "msdata")
+#' library(MsDataHub)
+#' fl <- MsDataHub::PestMix1_DDA.mzML()
 #'
 #' dda <- readMsExperiment(fl)
 #'
@@ -1711,13 +1712,13 @@ setGeneric("loadRaw", function(object, ...) standardGeneric("loadRaw"))
 #' @examples
 #'
 #' ## Read a test dataset.
-#' fls <- c(system.file("microtofq/MM14.mzML", package = "msdata"),
-#'          system.file("microtofq/MM8.mzML", package = "msdata"))
+#' library(MsDataHub)
+#' fls <- MsDataHub::PestMix1_DDA.mzML()
 #'
 #' ## Define a data frame with some sample annotations
 #' ann <- data.frame(
-#'     injection_index = 1:2,
-#'     sample_id = c("MM14", "MM8"))
+#'     injection_index = 1,
+#'     sample_id = c("Pest_mix"))
 #'
 #' ## Import the data
 #' library(MsExperiment)