+++ longtests/testthat/test_methods-OnDiskMSnExp.R
@@ -11,11 +11,6 @@ test_that("profMat,OnDiskMSnExp works", {
     res_2 <- profMat(xcmsRaw(faahko_3_files[2], profstep = 0), step = 2,
                      method = "binlin")
     expect_equal(res_2, res[[1]])
-
-    ## Simulating issue #312
-    od_1 <- filterFile(microtofq_od, 1)
-    od_1_clnd <- clean(removePeaks(od_1, t = 1800))
-    res_clnd <- profMat(od_1_clnd)
 })
 
 test_that("findChromPeaks,OnDiskMSnExp,CentWaveParam works", {
@@ -102,7 +97,7 @@ test_that("findChromPeaks,OnDiskMSnExp,CentWavePredIsoParam works", {
 })
 
 test_that("findChromPeaks,OnDiskMSnExp,MassifquantParam works", {
-    mzf <- system.file("microtofq/MM14.mzML", package = "msdata")
+    mzf <- faahko_3_files[1L]
     mqp <- MassifquantParam(ppm = 20, criticalValue = 1.2)
     res <- xcmsSet(mzf[1], method = "massifquant", ppm = 20,
                    criticalValue = 1.2)