+++ tests/testthat/test_XcmsExperiment-functions.R
@@ -83,9 +83,7 @@ test_that(".pmat_filter_mz works", {
 
 test_that(".chrom_peak_intensity_msw  works", {
     library(MsExperiment)
-    fticrf <- list.files(system.file("fticr-mzML", package = "msdata"),
-                         recursive = TRUE, full.names = TRUE)
-    fls <- normalizePath(fticrf)[1:2]
+    fls <- fticrf
     mp <- MSWParam(scales = c(1, 7), peakThr = 80000, ampTh = 0.005,
                    SNR.method = "data.mean", winSize.noise = 500)